+++ examples/complete/variables.tf
@@ -1,1 +1,5 @@
-#(empty)
+variable "name" {
+  description = "The name of the ECR repository"
+  type        = string
+  default     = "terraform-aws-ecr-repository"
+}